Design and implement a set of Java classes that implement the planner, the test harness and the GUI.
The planner itself, which implements a specified API.
A test harness that uses that API to test the behaviour of the
simulation.
A basic GUI that allows an ordinary university student (who is not a
programmer) to use the planner.
This must use the API, and only the API, to interact with the planner
code itself